backend/requirements.txt
# FastAPI e dependências web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Banco de dados
sqlalchemy==2.0.23
alembic==1.12.1

# IA e processamento
anthropic==0.7.8
openai==1.3.7

# Utilitários
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
jinja2==3.1.2
pyyaml==6.0.1
bleach==6.1.0

# Monitoramento de arquivos
watchdog==3.0.0

# Cache e filas
redis==5.0.1
celery==5.3.4

# Logs e métricas
structlog==23.2.0
prometheus-client==0.19.0

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Desenvolvimento
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Retry e resiliência
tenacity==8.2.3

# Validação e sanitização
email-validator==2.1.0 